Reimplement the TxNumber value-type to ensure that a slot of this type gets the correct value in memory once it is committed.  As it was, the correct value was being written to the persistent store, but the value in memory was always zero.
Unfortunately, to fix this, I had to change the programmers interface
to this value-type.  So, whereas before the code that was generated
transformed a TxNumber slot into a long, now it is actually an
instance of pt.ist.fenixframework.TxNumber, which is a new class that
contains the getNumber() method to access the actual number.