Several changes towards getting rid of the old idInternal and foreignKeys in the fenix-framework.
Changed the implementation of the FenixCache, merging in the code
written by Sergio Fernandes in the HBase branch and further
simplifying it.

WARNING for the Fenix webapp: in this version the
Transaction.readDomainObject method will always go to the database to
fetch the object, without looking first into the cache.  This does not
affect the correction of the code but may affect the performance.
Either way, the readDomainObject method should be eliminated, so this
is a step towards that end.

Still, if this change proves to be unacceptable for the Fenix webapp,
we may have to revert to the old two-levels FenixCache.