Drop all the support for the old idInternal-based foreign keys. Now, only the OIDs are used.
WARNING: If you have a working system that is using the
fenix-framework, make sure that your database has all the OIDs
correctly defined.  Otherwise, upgrading to this version of the
fenix-framework will not work with your current database.

Use the pt.ist.fenixframework.pstm.repository.CheckOids program to
check if the OIDs are OK.