Add new code generator that uses only a single box for storing the entire state of an object (other than collections that implement relations of multiplicity greater than 1).
This code generator also uses a lazy approach to the creation of
relation lists for storing the collections.

To use this new code generator, configure the
fenix.framework.codeGenerator property so that it has the value of
pt.ist.fenixframework.pstm.dml.FenixCodeGeneratorOneBoxPerObject,
instead of the default.

Changing the in-memory layout of the objects, as per this change, may
be switched back and forth, without any consequence on the database
schema or data.  So, you may experiment with this new layout and see
whether it fits better with your application.