[jvstm-common] Deal with a possible CommitException in the body of a transaction
This is only possible in JVSTM2.x due to an EarlyAbortException, which detects
an already invalid read during a write transaction.  In JVSTM1.x, this
exception is never thrown during the execution of the 'body' of a transaction,
so this change does not affect it.  Nevertheless, by doing this here, we're
able to reuse the transactional machinery implemented in jvstm-common.