[jvstm-common] Remove synchronized from reload() in VBox
It is not strictly necessary.  It only existed to prevent two reload requests
regarding the same vbox to hit the repository.  However, the actual required
synchronization is performed in the mergeVersions method.

Now that VBox is used also in the lock-free implementation, the extraneous
synchronized was removed so that it does not prevent the lock-free behavior.